+++ src/main/java/com/example/guerraalantico/Persistencia/DAO/EquipoBD.java
@@ -79,7 +79,7 @@ public class EquipoBD {
         NaveDTO naveDTO = new NaveDTO(rs.getInt("NavIdNave"), rs.getString("NavTipoNave"),
                 rs.getInt("RENCoordenadasX"), rs.getInt("RenCoordenadasY"),
                 rs.getInt("RENResistenciaActual"), rs.getInt("RenProfundidadActual"),
-                rs.getInt("NavVelocidad"), rs.getInt("RENRotacion"));
+                rs.getInt("NavVelocidad"), rs.getFloat("RENRotacion"));
         listaNaves.add(naveDTO);
       }
       rs.close();
